profile: Change terminfo to xterm-kitty
Kitty has its own exotic terminfo. In order to handle escape sequences
such as those for italics we need to use xterm-kitty which is a
strict requirement described in the documentation. This may have other
consequences but allows tmux and kitty to play nice with each other for
the most part.